文档整理及继电器指令生成程序在test/CRCmod
