+++ README.md
@@ -1,0 +1,11 @@
+# arm_project
+
+运行于linux系统的一套温室控制系统基站部分，根据温室环境自动或人工控制继电器，记录温室环境数据上传到服务器。
+
+## 基站 2.0 新功能
+1. 本地缓存 发送数据到服务器
+2. 数据库初始化
+
+## 服务器 2.0 新功能
+1. 接受基站发送的功能
+2. 数据库初始化
