--- control.py
@@ -1,257 +1,0 @@
-# coding=utf-8
-
-import json
-from currenttime import get_current_time
-from relay_output import tri_state_relay_output, bi_state_relay_output
-
-
-class Control(object):
-    '''the object of tri-state actuators and bi-state'''
-
-    def __init__(self):
-        self.__update_time = "2016/07/27 22:08:00"
-        self.__roof_vent_south = "stop"
-        self.__roof_vent_north = "stop"
-        self.__side_vent = "stop"
-        self.__shade_screen_north = "stop"
-        self.__shade_screen_south = "stop"
-        self.__thermal_screen = "stop"
-
-        self.__cooling_pump = "off"
-        self.__cooling_fan = "off"
-        self.__fan = "off"
-        self.__fogging = "off"
-        self.__heating = "off"
-        self.__co2 = "off"
-        self.__lighting_1 = "off"
-        self.__lighting_2 = "off"
-        self.__irrigation = "off"
-
-
-    def get_roof_vent_south(self):
-        return self.__roof_vent_south
-
-
-    def get_roof_vent_north(self):
-        return self.__roof_vent_north
-
-
-    def get_side_vent(self):
-        return self.__side_vent
-
-
-    def get_shade_screen_north(self):
-        return self.__shade_screen_north
-
-
-    def get_shade_screen_south(self):
-        return self.__shade_screen_south
-
-
-    def get_thermal_screen(self):
-        return self.__thermal_screen
-
-
-    def get_cooling_pump(self):
-        return self.__cooling_pump
-
-
-    def get_cooling_fan(self):
-        return self.__cooling_fan
-
-
-    def get_fan(self):
-        return self.__fan
-
-
-    def get_fogging(self):
-        return self.__fogging
-
-
-    def get_heating(self):
-        return self.__heating
-
-
-    def get_co2(self):
-        return self.__co2
-
-
-    def get_lighting_1(self):
-        return self.__lighting_1
-
-
-    def get_lighting_2(self):
-        return self.__lighting_2
-
-    def get_irrigation(self):
-        return self.__irrigation
-
-    def get_update_time(self):
-        return self.__update_time
-
-    def set_update_time(self, value):
-        self.__update_time = value
-
-    def set_roof_vent_south(self, value):
-        self.__roof_vent_south = value
-
-
-    def set_roof_vent_north(self, value):
-        self.__roof_vent_north = value
-
-
-    def set_side_vent(self, value):
-        self.__side_vent = value
-
-
-    def set_shade_screen_north(self, value):
-        self.__shade_screen_north = value
-
-
-    def set_shade_screen_south(self, value):
-        self.__shade_screen_south = value
-
-
-    def set_thermal_screen(self, value):
-        self.__thermal_screen = value
-
-
-    def set_cooling_pump(self, value):
-        self.__cooling_pump = value
-
-
-    def set_cooling_fan(self, value):
-        self.__cooling_fan = value
-
-
-    def set_fan(self, value):
-        self.__fan = value
-
-
-    def set_fogging(self, value):
-        self.__fogging = value
-
-
-    def set_heating(self, value):
-        self.__heating = value
-
-
-    def set_co2(self, value):
-        self.__co2 = value
-
-
-    def set_lighting_1(self, value):
-        self.__lighting_1 = value
-
-    def set_lighting_2(self, value):
-        self.__lighting_2 = value
-
-    def set_irrigation(self, value):
-        self.__irrigation = value
-
-    def handle_post(self, data):
-        obj = json.loads(data)
-        keys = obj.keys()
-        json_response = "{"
-        for key in keys:
-            if key == 'shade_screen_out':
-                key = 'shade_screen_south'
-            if key == 'shade_screen_in':
-                key = 'shade_screen_north'
-            if key == 'cooling_pad':
-                key = 'cooling_pump'
-            if key in Control.tri_states_actuators:
-                if key == 'shade_screen_south':
-                    value = obj.get('shade_screen_out')
-                elif key == 'shade_screen_north':
-                    value = obj.get('shade_screen_in')
-                else:
-                    value = obj.get(key)
-                if value in Control.tri_states:
-                    setattr(self, "_Control__" + key, value)
-                    print key, getattr(self, "_Control__" + key)
-                    try:
-                        tri_state_relay_output(key, value)
-                    except:
-                        print "the relay don't have the actuator, please review"
-                    json_response += '''"%s" : "%s",''' % (key, value)
-                else:
-                    print value, "tri illegal state"
-            elif key in Control.bi_states_actuators:
-                if key == 'cooling_pump':
-                    value = obj.get('cooling_pad')
-                else:
-                    value = obj.get(key)
-                if value in Control.bi_states:
-                    setattr(self, "_Control__" + key, value)
-                    print key, getattr(self, "_Control__" + key)
-                    try:
-                        bi_state_relay_output(key, value)
-                    except:
-                        print "the relay don't have the actuator, please review"
-                    json_response += '''"%s" : "%s", ''' % (key, value)
-                else:
-                    print value, "illegal state"
-            else:
-                print key, "bi illegal actuator"
-        json_response += '''"status" : "%s", ''' % "success"
-        json_response += '''"update_time" : "%s"''' % get_current_time()
-        json_response += "}"
-        return json_response
-
-    def build_json(self):
-        return '''
-        {
-      "update_time": "%s",
-      "actuator": {
-        "tri_state": {
-            "roof_vent_south": "%s",
-            "roof_vent_north": "%s",
-            "side_vent": "%s",
-            "shade_screen_north": "%s",
-            "shade_screen_south": "%s",
-            "thermal_screen": "%s"
-        },
-        "bi_state": {
-            "cooling_pump": "%s",
-            "cooling_fan":"%s",
-            "fan":"%s",
-            "fogging":"%s",
-            "heating": "%s",
-            "co2": "%s",
-            "lighting_1": "%s",
-            "lighting_2": "%s",
-            "irrigation": "%s"
-                }
-            }
-        }''' \
-               % ( self.__update_time,
-                  self.__roof_vent_south,
-                  self.__roof_vent_north,
-                  self.__side_vent,
-                  self.__shade_screen_north,
-                  self.__shade_screen_south,
-                  self.__thermal_screen,
-                  self.__cooling_pump,
-                  self.__cooling_fan,
-                  self.__fan,
-                  self.__fogging,
-                  self.__heating,
-                  self.__co2,
-                  self.__lighting_1,
-                  self.__lighting_2,
-                  self.__irrigation
-        )
-
-    tri_states_actuators = ("roof_vent_south", "roof_vent_north", "side_vent",
-                            "shade_screen_north", "shade_screen_south", "thermal_screen")
-
-    bi_states_actuators = ("cooling_pump","cooling_fan ","fan", "fogging", "heating", "co2", "lighting_1", "lighting_2", "irrigation")
-
-    tri_states = ("on", "off", "stop")
-
-    bi_states = ("on", "off")
-
-if __name__=='__main__':
-    c=Control()
-    print c.build_json()
