+++ docs/problems.md
@@ -1,0 +1,16 @@
+# 目前项目待解决问题
+
+## 基站 arm_project
+1. 定义宏 继电器与控制机构 修改执行机构名称的时候，需要修改 control.py relay_out.py
+
+## Java UI 界面
+1. 时间不对
+2. 判断与 服务器 连接状态
+3. 参数设置最后一排丢失（触控板）
+
+## BBB 问题
+1. linux系统时间设置 
+2. 网络连接固定IP
+3. 双击启动main.py 最小化进程不显示了 （linux启动器）
+
+
