+++ docs/readme.txt
@@ -1,15 +1,5 @@
 
 
-1、侧窗等待时间的计数以及单独运行具体放在哪一个模块还没有想好
-
-2、有关光照控制算法的运行是需要单独运行的，这一部分也没有处理好
-
-3、算法中的延时操作好了，三态中开启一段时间后，均要置位stop
-
-4、还剩下串口输出部分和  控制方式的选择等。。。。
-
-。。。。。没想通。。
-
 
 1、多线程调试基本上没有问题，现在还有问题是短时间是同时输出的，这个还需要待定。。
