+++ docs/室内气候状态indoor.txt
