+++ docs/室外气候状态outdoor.txt
