+++ docs/执行机构的控制状态.txt
