+++ test/test02.py
@@ -3,11 +3,12 @@
 @author: Zxh
 '''
 
-from outdoor import Outdoor
-from control import Control
+from currenttime import get_current_month, get_current_time, get_time
 from indoor import Indoor
+from outdoor import Outdoor
 from parameter import Parameter
-from currenttime import get_current_hour, get_current_month, get_current_time, get_time
+
+from arm_project.control import Control
 
 
 class Ghauto:
